修改统计页面打卡率取整的问题

+++ app/src/main/java/com/beitone/signup/ui/home/StatisticsFragment.java
@@ -222,7 +222,7 @@ public class StatisticsFragment extends BaseHomeFragment {
         if (rate.contains("%")) {
             rate = rate.replace("%", "");
         }
-        progressRateView.setRate((int) Double.parseDouble(rate));
+        progressRateView.setRate(Double.parseDouble(rate));
     }
 
     @Override
